+++ js/game/const.js
@@ -1,0 +1,13 @@
+const rootEl = document.querySelector('.arenas');
+
+const HIT = {
+  head: 30,
+  body: 25,
+  foot: 20,
+}
+
+const ATTACK = ['head', 'body', 'foot'];
+
+const form = document.querySelector('.control');
+
+export { rootEl, HIT, ATTACK, form }
